OPI conformance audit: fix BRC-20 protocol, controller, precompiles, and self-mint support
- Fix u64→u128 amounts, case normalization, ticker validation, zero rejection
- Embed real 8KB BRC20_Controller.bin bytecode replacing 5-byte stub
- Correct function selectors (mint/burn/balanceOf use bytes, not string)
- Fix production bug: missing evm.database(MetashrewDB) in view.rs
- Implement locked pkscript precompile with full Taproot P2TR construction
- Add self-mint support (5-byte tickers at block 837,090)
- Add 6-byte predeploy tickers (at block 912,690, alphanumeric/dash only)
- Add 301 tests: conformance, controller, precompiles, self-mint, OPI

+++ crates/shrew-test-helpers/src/assertions.rs
@@ -34,7 +34,7 @@ pub fn assert_inscription_count(expected: u32) {
 }
 
 /// Assert BRC20 balance for an owner+ticker
-pub fn assert_brc20_balance(owner: &str, ticker: &str, expected_available: u64, expected_total: u64) {
+pub fn assert_brc20_balance(owner: &str, ticker: &str, expected_available: u128, expected_total: u128) {
     let table = Brc20Balances::new();
     let data = table.get(owner, ticker).expect(&format!("No BRC20 balance for {}:{}", owner, ticker));
     let balance: Balance = serde_json::from_slice(&data).expect("Failed to deserialize balance");
@@ -45,7 +45,7 @@ pub fn assert_brc20_balance(owner: &str, ticker: &str, expected_available: u64, 
 }
 
 /// Assert BRC20 ticker supply
-pub fn assert_brc20_supply(ticker: &str, expected_supply: u64) {
+pub fn assert_brc20_supply(ticker: &str, expected_supply: u128) {
     let table = Brc20Tickers::new();
     let data = table.get(ticker).expect(&format!("BRC20 ticker {} not found", ticker));
     let entry: Ticker = serde_json::from_slice(&data).expect("Failed to deserialize ticker");
